Move run_output_thread() of browser.py to protocol.py
